[FIXED] single dispatcher : if an argument is nillable, it can accept nil value
