+++ elenasrc3/elc/cliconst.h
@@ -13,7 +13,7 @@
 
 namespace elena_lang
 {
-   #define ELC_REVISION_NUMBER               0x0028
+   #define ELC_REVISION_NUMBER               0x0029
 
 #if defined _M_IX86 || _M_X64
