+++ elenasrc3/elc/compiler.h
@@ -1538,7 +1538,7 @@ namespace elena_lang
 
       void declareParameter(MethodScope& scope, SyntaxNode node, bool withoutWeakMessages, 
          bool declarationMode, bool& variadicMode, bool& weakSignature, bool& noSignature,
-         pos_t& paramCount, ref_t* signature, size_t& signatureLen);
+         pos_t& paramCount, ref_t* signature, size_t& signatureLen, bool& nillable);
 
       ref_t declareClosureParameters(MethodScope& methodScope, SyntaxNode argNode);
 
@@ -1703,7 +1703,7 @@ namespace elena_lang
 
       void injectVirtualCode(SyntaxNode classNode, ClassScope& scope, bool interfaceBased);
       void injectVirtualMultimethod(SyntaxNode classNode, SyntaxKey methodType, Scope& scope, 
-         ref_t targetRef, ClassInfo& info, mssg_t multiMethod);
+         ref_t targetRef, ClassInfo& info, mssg_t multiMethod, int nillableArgs);
       void injectVirtualEmbeddableWrapper(SyntaxNode classNode, SyntaxKey methodType,
          ref_t targetRef, ClassInfo& info, mssg_t multiMethod, bool abstractOne);
 
@@ -1713,12 +1713,12 @@ namespace elena_lang
       void injectInitializer(SyntaxNode classNode, SyntaxKey methodType, mssg_t message);
 
       bool injectVirtualStrongTypedMultimethod(SyntaxNode classNode, SyntaxKey methodType, Scope& scope, 
-         mssg_t message, mssg_t resendMessage, TypeInfo outputInfo, Visibility visibility, bool isExtension);
+         mssg_t message, mssg_t resendMessage, TypeInfo outputInfo, Visibility visibility, bool isExtension, int nillableArgs);
       bool injectVirtualStrongTypedVariadicMultimethod(SyntaxNode classNode, SyntaxKey methodType, ModuleScopeBase& scope,
          mssg_t message, mssg_t resendMessage, ref_t outputRef, Visibility visibility, bool isExtension);
 
       void injectVirtualMultimethod(SyntaxNode classNode, SyntaxKey methodType, Scope& scope, 
-         ref_t targetRef, ClassInfo& classInfo, mssg_t message, bool inherited, TypeInfo outputInfo, Visibility visibility);
+         ref_t targetRef, ClassInfo& classInfo, mssg_t message, bool inherited, TypeInfo outputInfo, Visibility visibility, int nillableArgs);
       void injectVirtualMultimethod(SyntaxNode classNode, SyntaxKey methodType, Scope& scope, mssg_t message,
          mssg_t resendMessage, ref_t resendTarget, TypeInfo outputInfo, Visibility visibility, bool isExtension);
